MXS-4700: Treat GTID sequences as 64-bit integers
The sequence is a 64-bit integer and should be treated as
such. Otherwise the comparisons are not guaranteed to be correct.

+++ server/modules/routing/pinloki/find_gtid.cc
@@ -167,7 +167,7 @@ bool search_file(const std::string& file_name,
 
             maxsql::GtidListEvent event = rpl.gtid_list();
 
-            uint32_t highest_seq = 0;
+            uint64_t highest_seq = 0;
             bool domain_in_list = false;
 
             for (const auto& tid : event.gtid_list.gtids())
